Changed allocator to boost::fast_pool_allocator and reduced sizeof(Order) by half (#10)
* Added more tests

* Changed allocator to boost::fast_pool_allocator and reduced sizeof(Order) by half

+++ include/BaseTypes.hpp
@@ -13,6 +13,6 @@ enum class TradeSide : std::uint8_t {
     Sell = 'S'
 };
 
-using Price = std::int64_t;
-using Quantity = std::uint64_t;
+using Price = std::int32_t;
+using Quantity = std::uint32_t;
 using OrderId = std::uint64_t;
